fix: decimal a float64

+++ src/main/java/com/olc1/visitor/interpreter/value/DecimalValue.java
@@ -1,10 +1,10 @@
 package com.olc1.visitor.interpreter.value;
 
 public record DecimalValue(double value, int line, int column) implements ValueWrapper {
-    
+
     @Override
     public String getTypeName() {
-        return "decimal";
+        return "float64";
     }
 
     @Override
